[Feature] Specify resource requests and limits on profiling container (#34)
Users can specify container resource and limits with the following flags:
`cpu.requests`, `cpu.limits`, `mem.requests`, `mem.limits`. All these
flags are optional.

In order to allow graceful modification of the Job spec, a new type is
added: `JobDetails`, which can be used to pass configuration options
from the CLI to the Job spec. Right now only the `Resources` field of the
Job API is targeted, but this can be extended in the future.

In order to keep the list of function arguments small, a new container
type `FlameConfig` is introduced which holds `TargetDetails`, `JobDetails`
and CLI options. All functions are passed a pointer to this config.

Also extends error handling, adds comments and tests.